yash.core.azure.deservices.ingestion/pyproject.toml
[tool.poetry]
name = "yash.core.azure.deservices.ingestion"
version = "1.0.0"
description = "Ingestion repository that will be shared across projects"
authors = ["DA-Data Engineering"]
readme = "README.md"
repository = "https://dev.azure.com/YashCPTG/CPTG/_git/yash.core.azure.deservices.ingestion"
classifiers = [
    "Environment :: Console",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Operating System :: POSIX",
    "Operating System :: Microsoft :: Windows :: Windows 10",
    "License :: Other/Proprietary License"
]
packages = [
    { include = "yash/core/azure/data/engineering/helpers", from = "src" }
]
include = [
    "src/yash/core/azure/data/engineering/helpers/",
]

[tool.poetry.dependencies]
pytest = "^7.2.1"
python = "^3.8"
click = "^8.1.3"
pydantic = "^1.10.2"
msal = "^1.20.0"
junitparser = "^2.8.0"
xmlrunner = "^1.7.7"
unittest-xml-reporting = "^3.2.0"
PyYAML= "^6.0.1"

#[tool.poetry.group.tests]
#optional = true
[tool.poetry.group.test.dependencies]
pytest = "^7.2.1"
pytest-cov = "^4.0.0"
pandas = "^1.5.1"
azure-storage-file-datalake = "^12.9.1"
azure-keyvault-secrets = "^4.6.0"
azure-identity = "^1.11.0"
azure-mgmt-datafactory = "^3.1.0"

#[tool.poetry.group.dev]
#optional = true
[tool.poetry.group.dev.dependencies]
pytest = "^7.2.1"
flake8 = "^5.0.4"
flake8-junit-report = "^2.1.0"
pre-commit = "*"

[tool.pytest.ini_options]
testpaths = ["tests"]
filterwarnings = [
    "ignore::FutureWarning",
    "ignore::RuntimeWarning",
    "ignore::DeprecationWarning",
    "ignore::PendingDeprecationWarning",
    "ignore:PytestUnraisableExceptionWarning",
]

[tool.coverage.run]
source = ["."]
branch = false
omit = [
    "*.ci*",
    "*devops*",
    "*tests*",
    "*docs*",
    "setup.py"
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "raise AssertionError",
    "if __name__ == '__main__':",
    "declare_namespace",
]
ignore_errors = false
skip_covered = true
omit = [
    "*.ci*",
    "*devops*",
    "*tests*",
    "*docs*",
    "setup.py"
]
show_missing = true
